let renderer handle more common commands

+++ cocos/renderer/backend/Buffer.h
@@ -9,10 +9,10 @@ CC_BACKEND_BEGIN
 class Buffer : public cocos2d::Ref
 {
 public:
-    virtual void updateData(void* data, uint32_t size) = 0;
+    virtual void updateData(void* data, size_t size) = 0;
     
 protected:
-    Buffer(uint32_t size, BufferType type, BufferUsage usage)
+    Buffer(size_t size, BufferType type, BufferUsage usage)
     : _size(size)
     , _type(type)
     , _usage(usage)
@@ -22,7 +22,7 @@ protected:
     
     BufferUsage _usage = BufferUsage::READ;
     BufferType _type = BufferType::VERTEX;
-    uint32_t _size = 0;
+    size_t _size = 0;
 };
 
 CC_BACKEND_END
