+++ cocos/renderer/backend/metal/BufferMTL.h
@@ -8,10 +8,10 @@ CC_BACKEND_BEGIN
 class BufferMTL : public Buffer
 {
 public:
-    BufferMTL(id<MTLDevice> mtlDevice, uint32_t size, BufferType type, BufferUsage usage);
+    BufferMTL(id<MTLDevice> mtlDevice, size_t size, BufferType type, BufferUsage usage);
     ~BufferMTL();
     
-    virtual void updateData(void* data, uint32_t size) override;
+    virtual void updateData(void* data, size_t size) override;
     
     inline id<MTLBuffer> getMTLBuffer() const { return _mtlBuffer; }
