+++ cocos/renderer/backend/metal/BufferMTL.mm
@@ -2,7 +2,7 @@
 
 CC_BACKEND_BEGIN
 
-BufferMTL::BufferMTL(id<MTLDevice> mtlDevice, uint32_t size, BufferType type, BufferUsage usage)
+BufferMTL::BufferMTL(id<MTLDevice> mtlDevice, size_t size, BufferType type, BufferUsage usage)
 : Buffer(size, type, usage)
 {
     _mtlBuffer = [mtlDevice newBufferWithLength:size options:MTLResourceStorageModeShared];
@@ -13,7 +13,7 @@ BufferMTL::~BufferMTL()
     [_mtlBuffer release];
 }
 
-void BufferMTL::updateData(void* data, uint32_t size)
+void BufferMTL::updateData(void* data, size_t size)
 {
     memcpy(_mtlBuffer.contents, data, size);
 }
